Harmonise the types of the App's css path declarations
The CSS path for an app can be given either as an init parameter, or as a
class variable. The declared types for these two declarations didn't
match (the class variable didn't allow for a PurePath).

This ensures that the types are the same and any future changes will be made
to both routes of declaration.

See #866.