Change the way that `textual run` does the argv monkey-patch
A `textual run` already made an attempt to monkey-patch `argv` so that the
application being run would not see an `argv` any differently to how it
would if it had been run "vanilla". However, it could be caught out by code
like:

    from sys import argv

    ...

    do_something_with( argv )

beforehand it was okay with:

    import sys

    ...

    do_something_with( sys.argv )

With this change it should work both ways.

Somewhat related to #1064.