Let `textual run` run files that don't end in .py
It's not uncommon to create "standalone" commands in Python that don't end
in a .py and which:

  #!/usr/bin/ev python

or a variation on that theme. This commit makes it so that `textual run`
will treat a file that's named and set to run like that as if it were a file
that ends in .py.

See #1106.