Process the label on construction of a TreeNode
Currently there's an asymmetry in how the label is handled for a TreeNode.
If a str label is passed to the constructor it stays as a str type. On the
other hand, if it's set via set_label, it gets processed into a Rich Text
type. This commit removes that asymmetry.