Fix attribute type hinting and docstrings.
The type hint alone was not enough for the tooling to pick up the type, so we decided to add explicit type hints when assigning the attributes in __init__ because that is enough for the tooling, and having the explicit type hints in the code means type checkers can make sure the type hints add up. If we have explicit type hints in the docstrings, those can become outdated without us noticing.