Add some initial Input key/action unit tests
This is just getting a feel for how I'll go about testing these. The main
focus here won't be on the bindings themselves -- they're not really
interesting and I feel could change over time anyway as people's tastes
settle down. What I want to test here are the actions that get bound.

This is just an initial small set of what's going to be a much bigger
collection of Input action tests.