Remove the dictionary example from the front page of the docs
Sometimes, when building the docs, this would end up being a bit of an
"empty" example in that the time taken to get the result back from the API
would be so long that the output would be of the request just in progress.

So we've decided to drop this from the front page.