Bring the docs deployment target into the new approach
This one's a little vague right now, and I'm not sure how best to test this.
Looking at the mkdocs documentation it isn't clear to me that you *can*
provide the config file to the gh-build command, but there's example command
lines in the docs showing that you can.

Need to see if there's a safe way to test this out -- the last thing I want
to do is nuke our website.