Add some initial component classes
This might be enough actually, still need to think on this some more. There
is another issue to think though here too. I'm leaning heavily on a base
"toggle" button class from which the checkbox and the radio button derive.
This makes a lot of sense -- they're so similar that it doesn't really make
sense to do it another way.

But... component classes lean on a common prefix; and I can't see an obvious
way of declaring component classes based off the owning class name; not at
the moment anyway. Perhaps I can make that work.