Add all the mkdocs config files to the reload-watcher
When I made the changes to documentation generation I split up the then
single mkdocs configuration file. It looks like the reloader for mkdocs
always watches its configuration file, bit *only* the one that was loaded in
the first place, not any that are inherited from.

So this change adds all of the mkdocs yml files to the watch list to ensure
that any change causes the documentation server to reload.

Most useful if you're adding new files to the navigation.