Override can_view in Screen to check if a widget is overlay screen
This change is based on the idea that if a widget is "overlay: screen" it's
totally outside the normal flow of things and so trying to see if it's
visually within its ancestors doesn't make a whole lot of sense as it might
deliberately not be. At which point the only sensible question we can ask is
"can the screen see it?".

See #2958.