Add an initial take on getting a command to actually run
This is far from its final form, and right now to work it needs that the
calling code (which is in my test harness) receive the callable via the
screen callback system and make use of it. That's fine, it works, it just
means that as I get closer to making this part of Textual proper I'll need
to build such a mechanism into Screen.