Update to take into account very fast command sources
The code as was worked fine for nicely slow command sources. In fact I
thought it was going to be the slow ones that would give me the worst
problem. But having managed to handle that, it was the really fast ones that
showed a few issues.

Here I'm swapping back to just showing/hiding the loading indicator as
mounting and removing so fast was an issue.

And also I'm making sure that I flush the queue after all the tasks are
finished, otherwise it was easy to lose a lot of commands.