Allow "click and go" vs "click, review, then go"
During 2023-08-03 standup it was suggested to me that the preference might
be to have this work in a "click and go" way rather than a "click, review
then go" way. My preference is more for the latter, but I can see the desire
for the former too.

So this makes the former the default, but allows for the latter to be
configured by the developer.