Add initial basic app-wide support for the command palette
This adds an action to the App, and binds it to Ctrl+@, which in Mac
terminals at least equates to ctrl+space. This might not be the final
resting place for this, I'm not even sure if we should find it by default at
all. But for the purposes of further testing as I develop this that's fine.

This also adds support to the App class for running the user's choice of
command. With this change nobody needs to hook up the command palette with
Textual on their own any more, it's "out of the box".

What's needed by them now is hooking up a command provider.