Wait for the command sources to complete after input, before moving on
Just in the problematic tests. As an experiment for the moment. I've still
not quite got to the bottom of the core problem, as I've been seeing, but
there is an issue with testing the command palette: how to ensure that
there's actually matched commands before going on to test interaction, when
the sourcing of command matches is concurrent with anything else.

Here I reach in into the workers of the command palette and wait for them to
finish and *then* I go on to use the result.

Having been able to recreate the surface error locally, on a nice fast M2Pro
Mac, with a significant async sleep in the source, this fixed the error in
that situation. So let's see if that makes a dent in CI...