Drop the automatic generation of the FAQ (for now)
For this to work faqtory needs to be a development dependency of Textual.
Textual still maintains support for Python 3.7; faqtory is Python 3.8 or
greater. So, for the moment, we're going to cheat a little and make it so
that you have to remember to run faqtory to rebuild FAQ.md.

All hail the walrus!