Work harder to cancel running command search tasks
This commit takes the handling of running command search tasks a wee bit
further, sending "down" the aborted status and cancelling the tasks as soon
as possible. There are still situations where this won't really make a
difference, and depending on how the command source is coded it could carry
on running for a while, but if a command source is coded to handle being
cancelled as soon as possible this will provide what's needed to benefit
from such an approach.

Note that it *does* mean that a developer writing a command source, which
awaits something, will need to handle a CancelledError; we should probably
see about talking about this in the docs.