Avoid the found commands list flashing while typing
For obvious reasons, every time the user typed a letter, the list of
already-completed commands needed to be cleared down before new ones got
added. While the code concerned with doing this was in the right place (when
a key was pressed), this had the unfortunate side-effect of making the list
appear to "flash" as the user typed the first few letters, especially if a
lot of hits were found near-instantly.

This commit delays the initial clearing-down of the content of the list,
keeping track of if the clear has been done already and only doing it at the
very last moment if it's needed.