Squish the typing issue with asend
The asend back into the search routing was always showing a typing mismatch,
but I couldn't quite see what was going on; what made it even more confusing
was the code was working fine.

It looks like keeping hold of the "routine", and keeping that distinct from
the iterator of the results, is the trick here. It all still works *and* the
typing works out.