Remove the "no matches" note as soon as possible
We've recently changed the way the command list is cleared down when the
search term is modified, thus removing a source of "flashing" as the user
types; this pretty much involves *not* clearing down the previous hits until
the first new hit comes in. This is fine in all situations expect where the
last search was a "no matches" search.

In that situation the next search stats out saying "no matches". That's
correct, that's the result of the previous search, but in this case it's
unhelpful and potentially confusing. So this commit checks if that's the
state of the command list up front and clears that option from the list.